feat: バトル進行処理を新規作成し、進行時の状態更新を実装
